last commit I hope

+++ server/src/commands/RemoveCommand.java
@@ -1,12 +1,15 @@
 package commands;
 
 import exceptions.NonexistentKeyException;
+import models.MusicBand;
 import network.ResponseBuilder;
 import network.db.DmlQueryManager;
-import stateManager.CollectionManager;
+import manager.stateManager.CollectionManager;
+import network.db.UserManager;
 import utility.Program;
 
 import java.sql.SQLException;
+import java.util.Hashtable;
 
 public class RemoveCommand extends Command {
     public RemoveCommand() {
@@ -20,17 +23,23 @@ public class RemoveCommand extends Command {
      */
     @Override
     public void execute(String key) {
+        String username = user.getUsername();
+        Hashtable<String, MusicBand> collection = CollectionManager.getInstance().getCollection();
         ResponseBuilder responseBuilder = Program.getInstance().getResponseBuilder();
         try {
             if (!CollectionManager.getInstance().hasSuchKey(key)) {
                 throw new NonexistentKeyException(key);
             }
+            if (collection.get(key).getUser_id() != new UserManager(user).getId()) {
+                responseBuilder.add(username, "Вы не можете удалить объект, который вы не создавали.");
+                return;
+            }
             new DmlQueryManager(user).deleteMusicByKey(key);
-            responseBuilder.add("Элемент с ключом %s удален успешно.".formatted(key));
+            responseBuilder.add(username, "Элемент с ключом %s удален успешно.".formatted(key));
         } catch (NonexistentKeyException e) {
-            responseBuilder.add(e.getMessage());
+            responseBuilder.add(username, e.getMessage());
         } catch (SQLException e) {
-            responseBuilder.add("Ошибка при удалении элемента из базы данных!");
+            responseBuilder.add(username, "Ошибка при удалении элемента из базы данных!");
         }
     }
 }
